Enum plasmo::AlignContent

pub enum AlignContent {
    Start,
    End,
    FlexStart,
    FlexEnd,
    Center,
    Stretch,
    SpaceBetween,
    SpaceEvenly,
    SpaceAround,
}

Sets the distribution of space between and around content items For Flexbox it controls alignment in the cross axis For Grid it controls alignment in the block axis

MDN

Variants

Start

Items are packed toward the start of the axis

End

Items are packed toward the end of the axis

FlexStart

Items are packed towards the flex-relative start of the axis.

For flex containers with flex_direction RowReverse or ColumnReverse this is equivalent to End. In all other cases it is equivalent to Start.

FlexEnd

Items are packed towards the flex-relative end of the axis.

For flex containers with flex_direction RowReverse or ColumnReverse this is equivalent to Start. In all other cases it is equivalent to End.

Center

Items are centered around the middle of the axis

Stretch

Items are stretched to fill the container

SpaceBetween

The first and last items are aligned flush with the edges of the container (no gap) The gap between items is distributed evenly.

SpaceEvenly

The gap between the first and last items is exactly THE SAME as the gap between items. The gaps are distributed evenly

SpaceAround

The gap between the first and last items is exactly HALF the gap between items. The gaps are distributed evenly in proportion to these ratios.