Enum plasmo::MaxTrackSizingFunction

pub enum MaxTrackSizingFunction {
    Fixed(LengthPercentage),
    MinContent,
    MaxContent,
    FitContent(LengthPercentage),
    Auto,
    Fraction(f32),
}

Maximum track sizing function

Specifies the maximum size of a grid track. A grid track will automatically size between it’s minimum and maximum size based on the size of it’s contents, the amount of available space, and the sizing constraint the grid is being size under. See https://developer.mozilla.org/en-US/docs/Web/CSS/grid-template-columns

Variants

Fixed(LengthPercentage)

Track maximum size should be a fixed points or percentage value

MinContent

Track maximum size should be content sized under a min-content constraint

MaxContent

Track maximum size should be content sized under a max-content constraint

FitContent(LengthPercentage)

Track maximum size should be sized according to the fit-content formula

Auto

Track maximum size should be automatically sized

Fraction(f32)

The dimension as a fraction of the total available grid space (fr units in CSS) Specified value is the numerator of the fraction. Denominator is the sum of all fraction specified in that grid dimension Spec: https://www.w3.org/TR/css3-grid-layout/#fr-unit

Implementations

impl MaxTrackSizingFunction

pub fn is_intrinsic(&self) -> bool

Returns true if the max track sizing function is MinContent, MaxContent, FitContent or Auto, else false.

pub fn is_max_content_alike(&self) -> bool

Returns true if the max track sizing function is MaxContent, FitContent or Auto else false. “In all cases, treat auto and fit-content() as max-content, except where specified otherwise for fit-content().” See: https://www.w3.org/TR/css-grid-1/#algo-terms

pub fn is_flexible(&self) -> bool

Returns true if the max track sizing function is Flex, else false.

pub fn definite_value(self, parent_size: Option<f32>) -> Option<f32>

Returns fixed point values directly. Attempts to resolve percentage values against the passed available_space and returns if this results in a concrete value (which it will if the available_space is Some). Otherwise returns None.

pub fn definite_limit(self, parent_size: Option<f32>) -> Option<f32>

Resolve the maximum size of the track as defined by either: - A fixed track sizing function - A percentage track sizing function (with definite available space) - A fit-content sizing function with fixed argument - A fit-content sizing function with percentage argument (with definite available space) All other kinds of track sizing function return None.

pub fn resolved_percentage_size(self, parent_size: f32) -> Option<f32>

Resolve percentage values against the passed parent_size, returning Some(value) Non-percentage values always return None.

pub fn uses_percentage(self) -> bool

Whether the track sizing functions depends on the size of the parent node