Struct plasmo::Taffy

pub struct Taffy { /* private fields */ }

A tree of UI Nodes, suitable for UI layout

Implementations

impl Taffy

pub fn new() -> Taffy

Creates a new Taffy

The default capacity of a Taffy is 16 nodes.

pub fn with_capacity(capacity: usize) -> Taffy

Creates a new Taffy that can store capacity nodes before reallocation

pub fn enable_rounding(&mut self)

Enable rounding of layout values. Rounding is enabled by default.

pub fn disable_rounding(&mut self)

Disable rounding of layout values. Rounding is enabled by default.

pub fn new_leaf(&mut self, layout: Style) -> Result<DefaultKey, TaffyError>

Creates and adds a new unattached leaf node to the tree, and returns the Node of the new node

pub fn new_leaf_with_measure( &mut self, layout: Style, measure: MeasureFunc ) -> Result<DefaultKey, TaffyError>

Creates and adds a new unattached leaf node to the tree, and returns the Node of the new node

Creates and adds a new leaf node with a supplied [MeasureFunc]

pub fn new_with_children( &mut self, layout: Style, children: &[DefaultKey] ) -> Result<DefaultKey, TaffyError>

Creates and adds a new node, which may have any number of children

pub fn clear(&mut self)

Drops all nodes in the tree

pub fn remove(&mut self, node: DefaultKey) -> Result<DefaultKey, TaffyError>

Remove a specific Node from the tree and drops it

Returns the id of the node removed.

pub fn set_measure( &mut self, node: DefaultKey, measure: Option<MeasureFunc> ) -> Result<(), TaffyError>

Sets the [MeasureFunc] of the associated node

pub fn add_child( &mut self, parent: DefaultKey, child: DefaultKey ) -> Result<(), TaffyError>

Adds a child Node under the supplied parent

pub fn set_children( &mut self, parent: DefaultKey, children: &[DefaultKey] ) -> Result<(), TaffyError>

Directly sets the children of the supplied parent

pub fn remove_child( &mut self, parent: DefaultKey, child: DefaultKey ) -> Result<DefaultKey, TaffyError>

Removes the child of the parent node

The child is not removed from the tree entirely, it is simply no longer attached to its previous parent.

pub fn remove_child_at_index( &mut self, parent: DefaultKey, child_index: usize ) -> Result<DefaultKey, TaffyError>

Removes the child at the given index from the parent

The child is not removed from the tree entirely, it is simply no longer attached to its previous parent.

pub fn replace_child_at_index( &mut self, parent: DefaultKey, child_index: usize, new_child: DefaultKey ) -> Result<DefaultKey, TaffyError>

Replaces the child at the given child_index from the parent node with the new child node

The child is not removed from the tree entirely, it is simply no longer attached to its previous parent.

pub fn child_at_index( &self, parent: DefaultKey, child_index: usize ) -> Result<DefaultKey, TaffyError>

Returns the child Node of the parent node at the provided child_index

pub fn total_node_count(&self) -> usize

Returns the total number of nodes in the tree

pub fn child_count(&self, parent: DefaultKey) -> Result<usize, TaffyError>

Returns the number of children of the parent Node

pub fn children( &self, parent: DefaultKey ) -> Result<Vec<DefaultKey>, TaffyError>

Returns a list of children that belong to the parent Node

pub fn set_style( &mut self, node: DefaultKey, style: Style ) -> Result<(), TaffyError>

Sets the Style of the provided node

pub fn style(&self, node: DefaultKey) -> Result<&Style, TaffyError>

Gets the Style of the provided node

pub fn layout(&self, node: DefaultKey) -> Result<&Layout, TaffyError>

Return this node layout relative to its parent

pub fn dirty(&self, node: DefaultKey) -> Result<bool, TaffyError>

Indicates whether the layout of this node (and its children) need to be recomputed

pub fn compute_layout( &mut self, node: DefaultKey, available_space: Size<AvailableSpace> ) -> Result<(), TaffyError>

Updates the stored layout of the provided node and its children